backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pdfplumber
openai
python-multipart
icalendar
